chore: rename folder calendar
